checkpoint push: single crawler works, but __main__ broken
