+++ pyproject.toml
@@ -17,7 +17,6 @@ cloudscraper = "1.2.71"
 colorama = "0.4.4"
 pillow = "10.2.0"
 pretty-errors = "1.2.19"
-requests = "2.31.0"
 tqdm = "4.59.0"
 # https://stackoverflow.com/questions/446209/possible-values-from-sys-platform
 pywin32 = {version = "^306", markers = "sys_platform == 'win32'"}
@@ -29,6 +28,8 @@ confz = "^2.0.1"
 pydantic-extra-types = "^2.9.0"
 pendulum = "^3.0.0"
 slimeface = "^2024.9.27"
+httpx = {extras = ["socks"], version = "^0.27.2"}
+aiofiles = "^24.1.0"
 
 [tool.poetry.scripts]
 javsp = "javsp.__main__:entry"
